some PR comments
